chore: 🤖 평접 타입 변경
✅ Closes: #S04P31A106-348

+++ backend/src/main/java/com/ssafy/checkmate/member/dto/Member.java
@@ -39,5 +39,5 @@ public class Member {
     private String memberProfileUrl;
 
     @NonNull
-    private int memberGrade;
+    private Double memberGrade;
 }
